# internal/swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Private bank API — syncs Fio Bank transaction data into Postgres, matches member payments to their bank transactions, and detects missed dues. Dev-only documentation (see config.EnableSwaggerDocs): every route below requires a Keycloak bearer token unless noted otherwise.",
        "title": "Bank System API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires the list-transactions role. All filters optional and combinable.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Browse processed transactions",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "true = has a member match, false = the unassigned worklist",
                        "name": "assigned",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "incoming or outgoing",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "e.g. membership_fee, salary, other_income, other_expense",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "variable_symbol, manual, or amount_heuristic",
                        "name": "matched_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "One member's transactions",
                        "name": "member_number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD, inclusive",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD, inclusive",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Default 100, capped at 500",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Default 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.transactionsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transactions/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires the view-budget role. Grouped by direction/category/currency only — no member_number or counterparty, safe for wide member-facing use.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Budgeting totals by category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD, inclusive",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD, inclusive",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.categorySummaryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires the list-transactions role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get one transaction's detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "processed_transactions ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.transactionDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.categorySummaryResponse": {
            "type": "object",
            "properties": {
                "incoming": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.categoryTotal"
                    }
                },
                "outgoing": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.categoryTotal"
                    }
                }
            }
        },
        "handler.categoryTotal": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "total": {
                    "type": "string"
                }
            }
        },
...
        "handler.monthRef": {
            "type": "object",
            "properties": {
                "month": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "handler.transactionDetail": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "constant_symbol": {
                    "type": "string"
                },
                "counter_account_name": {
                    "type": "string"
                },
                "counter_account_number": {
                    "type": "string"
                },
                "covered_months": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.monthRef"
                    }
                },
                "currency": {
                    "type": "string"
                },
                "direction": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_public_visible": {
                    "type": "boolean"
                },
                "matched_by": {
                    "type": "string"
                },
                "member_number": {
                    "type": "integer"
                },
                "message_for_recipient": {
                    "type": "string"
                },
                "specific_symbol": {
                    "type": "string"
                },
                "transaction_date": {
                    "type": "string"
                },
                "user_identification": {
                    "type": "string"
                },
                "variable_symbol": {
                    "type": "string"
                }
            }
        },
...
        "handler.transactionsResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.transactionListItem"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Keycloak-issued bearer token for bank-system's Keycloak client, sent as `Authorization: Bearer \u003ctoken\u003e`. Which role(s) a given route requires is noted on that route below — see internal/keycloak.Role for the full list.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}